1.0: fixed printf type position

+++ src/utils/diff.c
@@ -54,7 +54,7 @@ int diff(char* file1name, char *file2name, char *outfilename, int chunksize, int
 
     if(bytecnt1 != bytecnt2 || memcmp(buffer1, buffer2, chunksize) != 0) {
       if(verbose) {
-        printf("  DEBUG: Found difference at position %ld\n", position);
+        printf("  DEBUG: Found difference at position %ld\n", (long int)position);
       }
       // The files differ here. Write diff record.
       data_record_t diff_record;
@@ -62,7 +62,7 @@ int diff(char* file1name, char *file2name, char *outfilename, int chunksize, int
       diff_record.length = bytecnt2;
 
       if(verbose) {
-        printf("  DEBUG: Diff record: Position: %ld, length: %ld\n", diff_record.position, diff_record.length);
+        printf("  DEBUG: Diff record: Position: %ld, length: %ld\n", (long int)diff_record.position, (long int)diff_record.length);
       }
       fwrite(&diff_record.position, sizeof(diff_record.position), 1, outfile);
       fwrite(&diff_record.length,  sizeof(diff_record.length),  1, outfile);
